backend/pyproject.toml
[project]
name = "cppm-backend"
version = "0.1.0"
description = "CPPM Centralized Professional Profile Manager - Backend"
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "fastapi==0.135.1",
    "pydantic[email]==2.12.0",
    "pydantic-settings>=2.0.0",
    "alembic==1.18.4",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "httpx",
    "cryptography",
    "google-auth",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
    "bandit[toml]>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["app/database/alembic/*", "app/__init__.py", "app/**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
fail_under = 0

# --- Quality: Ruff (lint + format) ---
[tool.ruff]
target-version = "py314"
line-length = 100
src = ["app", "tests"]
exclude = ["app/database/alembic"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line length handled by formatter
    "B008",   # do not perform function calls in defaults — FastAPI Depends() is intentional
]

[tool.ruff.format]
quote-style = "double"

# --- Quality: Pyright (static types) ---
[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedVariable = true
# Exclude generated/boilerplate code; tighten once fixed
exclude = ["**/alembic/**"]

# --- Quality: Bandit (security) ---
[tool.bandit]
exclude_dirs = ["tests", "app/database/alembic"]
skips = [
    "B101",  # assert_used: allow in tests if needed
    "B105",  # hardcoded_password_string: OAuth URLs and literal "offline" are false positives
    "B110",  # try_except_pass: allow where intentionally silencing (e.g. optional token refresh)
]
